% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstplyr-package.R
\docType{package}
\name{fstplyr-package}
\alias{fstplyr}
\alias{fstplyr-package}
\title{fstplyr: A 'dplyr' Interface to 'fst'}
\description{
Provides a basic 'dplyr' interface to 'fst' data sources.
}
\seealso{
\code{\link[=src_fst]{src_fst()}}
}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
