% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/pipe.R, R/select.R, R/src.R,
%   R/tbl.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{reexports}
\alias{distinct}
\alias{reexports}
\alias{do}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{intersect}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{pull}
\alias{reexports}
\alias{sample_frac}
\alias{reexports}
\alias{sample_n}
\alias{reexports}
\alias{setdiff}
\alias{reexports}
\alias{setequal}
\alias{reexports}
\alias{slice}
\alias{reexports}
\alias{summarise}
\alias{reexports}
\alias{union}
\alias{reexports}
\alias{union_all}
\alias{reexports}
\alias{group_by}
\alias{reexports}
\alias{group_indices}
\alias{reexports}
\alias{group_size}
\alias{reexports}
\alias{groups}
\alias{reexports}
\alias{n_groups}
\alias{reexports}
\alias{ungroup}
\alias{reexports}
\alias{anti_join}
\alias{reexports}
\alias{full_join}
\alias{reexports}
\alias{inner_join}
\alias{reexports}
\alias{left_join}
\alias{reexports}
\alias{right_join}
\alias{reexports}
\alias{semi_join}
\alias{reexports}
\alias{as.tbl}
\alias{reexports}
\alias{as.tbl_cube}
\alias{reexports}
\alias{collapse}
\alias{reexports}
\alias{compute}
\alias{reexports}
\alias{same_src}
\alias{reexports}
\alias{tbl_vars}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{rename}
\alias{reexports}
\alias{src_tbls}
\alias{reexports}
\alias{tbl}
\alias{reexports}
\alias{collect}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{do}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{intersect}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{sample_frac}}, \code{\link[dplyr]{sample_n}}, \code{\link[dplyr]{setdiff}}, \code{\link[dplyr]{setequal}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{union}}, \code{\link[dplyr]{union_all}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{group_indices}}, \code{\link[dplyr]{group_size}}, \code{\link[dplyr]{groups}}, \code{\link[dplyr]{n_groups}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{anti_join}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{as.tbl}}, \code{\link[dplyr]{as.tbl_cube}}, \code{\link[dplyr]{collapse}}, \code{\link[dplyr]{compute}}, \code{\link[dplyr]{same_src}}, \code{\link[dplyr]{tbl_vars}}, \code{\link[dplyr]{\%>\%}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{src_tbls}}, \code{\link[dplyr]{tbl}}, \code{\link[dplyr]{collect}}}
}}

