% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{src_fst}
\alias{src_fst}
\title{Open a directory with .fst files as data source}
\usage{
src_fst(path)
}
\arguments{
\item{path}{An existing directory}
}
\description{
Create a data source to access all \code{.fst} files in a directory with
\pkg{dplyr} verbs.
}
\examples{
path <- tempfile()
dir.create(path)
fst::write_fst(iris, file.path(path, "iris.fst"))
fst::write_fst(mtcars, file.path(path, "mtcars"))

src <- src_fst(path)
src
tbl(src, "mtcars")
}
