% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fstplyr-package.R
\docType{package}
\name{fstplyr-package}
\alias{fstplyr}
\alias{fstplyr-package}
\title{fstplyr: A 'dplyr' Interface to 'fst'}
\description{
Provides a basic 'dplyr' interface to 'fst' data sources.
}
\details{
Work in progress.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/fstplyr}
  \item \url{http://krlmlr.github.io/fstplyr}
  \item Report bugs at \url{https://github.com/krlmlr/fstplyr/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
