% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/pipe.R, R/select.R, R/slice.R,
%   R/src.R, R/tbl.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{distinct}
\alias{do}
\alias{filter}
\alias{intersect}
\alias{mutate}
\alias{sample_frac}
\alias{sample_n}
\alias{setdiff}
\alias{setequal}
\alias{summarise}
\alias{union}
\alias{union_all}
\alias{group_vars}
\alias{group_by}
\alias{group_indices}
\alias{group_size}
\alias{groups}
\alias{n_groups}
\alias{ungroup}
\alias{anti_join}
\alias{full_join}
\alias{inner_join}
\alias{left_join}
\alias{right_join}
\alias{semi_join}
\alias{as_tibble}
\alias{as.tbl_cube}
\alias{collapse}
\alias{compute}
\alias{same_src}
\alias{tbl_vars}
\alias{\%>\%}
\alias{select}
\alias{rename}
\alias{pull}
\alias{slice}
\alias{src_tbls}
\alias{tbl}
\alias{collect}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cubelyr}{\code{\link[cubelyr]{as.tbl_cube}}}

  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr:filter-joins]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr:reexports]{as_tibble}}, \code{\link[dplyr:compute]{collapse}}, \code{\link[dplyr:compute]{collect}}, \code{\link[dplyr]{compute}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{do}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:mutate-joins]{full_join}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:group_data]{group_indices}}, \code{\link[dplyr:group_data]{group_size}}, \code{\link[dplyr:group_data]{group_vars}}, \code{\link[dplyr:group_data]{groups}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:reexports]{intersect}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr:group_data]{n_groups}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr:mutate-joins]{right_join}}, \code{\link[dplyr]{same_src}}, \code{\link[dplyr:sample_n]{sample_frac}}, \code{\link[dplyr]{sample_n}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:filter-joins]{semi_join}}, \code{\link[dplyr:reexports]{setdiff}}, \code{\link[dplyr:reexports]{setequal}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{src_tbls}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{tbl}}, \code{\link[dplyr]{tbl_vars}}, \code{\link[dplyr:group_by]{ungroup}}, \code{\link[dplyr:reexports]{union}}, \code{\link[dplyr:setops]{union_all}}}
}}

